Bug#128 Failure to detect chain using Olimex ARM-USB-OCD-H
Using the latest code, git revision 334e94, compiled and running under
Ubuntu 16.04, an Olimex ARM-USB-OCD-H adapter is correctly identified
by the "cable probe" command but fails to detect the jtag chain of a
connected STM32F4 device. The returned message indicates TDO is stuck
at 0.

The exact same hardware setup is working using OpenOCD git revision
(2b47de). In both cases the frequency is set to 1MHz. Using an
oscilloscope to monitor the signals, TCK and TDI have activity but TMS
and TDO remain low. It appears that TMS is not enabled properly which
could be a problem with initialization.

By following the initialization sequence, differences were observed
between OpenOCD and UrJTAG. We were able to get UrJTAG working with the
ARM-USB-OCD-H adapter by applying the following code changes. However,
this may not be optimal so hopefully someone with a better understanding
of the FT2232H can implement an appropriate correction.