Cable/J-Link: Dynamically search for usb endpoints
Between hardware revision 5 and 11, the USB endpoint ids were changed.
The old driver was hardcoded to specific endpoint ids. As a result,
newer J-Links simply did not work.

This patch dynamically searches for the correct endpoints. After that,
newer J-Links should work as expected.
The code for this was mostly copied from libjaylink's transport_usb.c

Signed-off-by: Stephan Brunner <s.brunner@stephan-brunner.net>